Move TestCircuit into own file and add bench feature
