+++ Makefile
@@ -5,4 +5,7 @@ fmt:
 	@cargo fmt
 
 clippy:
-	@cargo clippy --all-features -- -D warnings
+	@cargo clippy --all-features
+
+bench:
+	@cargo bench --features bench
