+++ src/lib.rs
@@ -1,10 +1,13 @@
 #![allow(clippy::too_many_arguments)]
-#![deny(unsafe_code)]
+#![deny(unsafe_code, unused_imports)]
 
+#[cfg(any(test, feature = "bench"))]
+mod circuit;
 pub mod constraint_builder;
 pub mod gadgets;
 mod mpt_table;
-pub mod tests;
+#[cfg(test)]
+mod tests;
 pub mod types;
 mod util;
 
@@ -14,3 +17,6 @@ pub mod serde;
 pub use gadgets::mpt_update::hash_traces;
 pub use mpt::MptCircuitConfig;
 pub use mpt_table::MPTProofType;
+
+#[cfg(feature = "bench")]
+pub use circuit::TestCircuit;
