+++ src/tests.rs
@@ -1,13 +1,9 @@
-use crate::{
-    gadgets::poseidon::PoseidonTable, hash_traces, serde::SMTTrace, types::Proof, MPTProofType,
-    MptCircuitConfig,
-};
+use crate::{circuit::TestCircuit, serde::SMTTrace, types::Proof, MPTProofType, MptCircuitConfig};
 use ethers_core::types::{Address, U256};
 use halo2_proofs::{
-    circuit::{Layouter, SimpleFloorPlanner},
     dev::MockProver,
     halo2curves::bn256::{Bn256, Fr},
-    plonk::{keygen_vk, Circuit, ConstraintSystem, Error, FirstPhase},
+    plonk::{keygen_vk, Circuit, ConstraintSystem},
     poly::kzg::commitment::ParamsKZG,
 };
 use mpt_zktrie::state::{builder::HASH_SCHEME_DONE, witness::WitnessGenerator, ZktrieState};
@@ -55,53 +51,6 @@ fn reverse(trace: SMTTrace) -> SMTTrace {
         update.reverse()
     }
     reversed
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct TestCircuit {
-    n_rows: usize,
-    proofs: Vec<Proof>,
-}
-
-impl TestCircuit {
-    pub fn new(n_rows: usize, traces: Vec<(MPTProofType, SMTTrace)>) -> Self {
-        Self {
-            n_rows,
-            proofs: traces.into_iter().map(Proof::from).collect(),
-        }
-    }
-}
-
-impl Circuit<Fr> for TestCircuit {
-    type Config = (PoseidonTable, MptCircuitConfig);
-    type FloorPlanner = SimpleFloorPlanner;
-
-    fn without_witnesses(&self) -> Self {
-        Self::default()
-    }
-
-    fn configure(cs: &mut ConstraintSystem<Fr>) -> Self::Config {
-        let poseidon = PoseidonTable::configure(cs);
-        let challenge = cs.challenge_usable_after(FirstPhase);
-        let mpt_circuit_config = MptCircuitConfig::configure(cs, challenge, &poseidon);
-        (poseidon, mpt_circuit_config)
-    }
-
-    fn synthesize(
-        &self,
-        config: Self::Config,
-        mut layouter: impl Layouter<Fr>,
-    ) -> Result<(), Error> {
-        let (poseidon, mpt_circuit_config) = config;
-        mpt_circuit_config.assign(&mut layouter, &self.proofs, self.n_rows)?;
-        layouter.assign_region(
-            || "load poseidon table",
-            |mut region| {
-                poseidon.load(&mut region, &hash_traces(&self.proofs));
-                Ok(())
-            },
-        )
-    }
 }
 
 fn mock_prove(witness: Vec<(MPTProofType, SMTTrace)>) {
